@@ -3,24 +3,32 @@
   "version": "0.2.0",
   "dependencies": {
     "@types/history": "^4.6.2",
+    "@types/isotope-layout": "^3.0.4",
+    "@types/jquery": "^3.3.1",
     "@types/react": "^16.0.40",
+    "@types/react-bootstrap": "^0.32.6",
     "@types/react-dom": "^16.0.4",
+    "@types/react-hot-loader": "3.0.3",
     "@types/react-redux": "^5.0.15",
     "@types/react-router-dom": "^4.2.5",
-    "@types/react-router-redux": "5.0.12",
-    "@types/redux": "3.6.0",
     "@types/webpack": "^4.1.0",
     "@types/webpack-env": "^1.13.5",
     "aspnet-prerendering": "^3.0.1",
+    "aspnet-webpack": "^2.0.3",
+    "aspnet-webpack-react": "^3.0.0",
     "awesome-typescript-loader": "^4.0.1",
     "babel-core": "^6.26.0",
     "babel-eslint": "^8.2.2",
     "babel-loader": "^7.1.4",
     "babel-plugin-module-resolver": "^3.0.0",
     "babel-plugin-transform-object-rest-spread": "^6.26.0",
+    "babel-polyfill": "^6.26.0",
     "babel-preset-env": "^1.6.1",
     "babel-preset-react": "^6.24.1",
-    "bootstrap": "^3.3.7",
+    "bootstrap": "^4.0.0",
+    "bootstrap-sass": "^3.3.7",
+    "compression-webpack-plugin": "^1.1.11",
+    "core-js": "^2.5.3",
     "create-react-class": "^15.6.3",
     "css-loader": "^0.28.11",
     "domain-task": "^3.0.0",
@@ -33,36 +41,49 @@
     "event-source-polyfill": "^0.0.12",
     "extract-text-webpack-plugin": "^4.0.0-beta.0",
     "file-loader": "^1.1.11",
+    "font-awesome": "^4.7.0",
+    "fontawesome": "^4.7.2",
     "history": "4.7.2",
+    "isomorphic-fetch": "^2.2.1",
     "isotope-layout": "^3.0.4",
     "jquery": "^3.3.1",
     "json-loader": "^0.5.4",
+    "kind-of": "^6.0.2",
+    "less": "^3.0.0",
     "less-loader": "^4.1.0",
     "module-resolver": "^1.0.0",
     "node-noop": "^1.0.0",
     "node-sass": "^4.5.3",
     "popper.js": "^1.14.1",
+    "postcss-loader": "^2.1.3",
     "prop-types": "^15.6.0",
+    "raf": "^3.4.0",
     "react": "^16.2.0",
-    "react-bootstrap": "^0.31.5",
+    "react-bootstrap": "^0.32.1",
     "react-dom": "^16.2.0",
-    "react-hot-loader": "^4.0.0",
+    "react-hot-loader": "^3.0.0-beta.7",
     "react-redux": "^5.0.6",
     "react-router": "^4.2.0",
+    "react-router-config": "^1.0.0-beta.4",
     "react-router-dom": "^4.2.2",
-    "react-router-redux": "^5.0.0-alpha.8",
+    "react-router-redux": "^5.0.0-alpha.9",
+    "react-static-routes": "^1.0.0",
     "react-time": "^4.3.0",
     "redux": "^3.7.2",
+    "redux-connect": "^7.0.0",
     "redux-thunk": "^2.2.0",
     "sass-loader": "^6.0.5",
     "style-loader": "^0.20.3",
     "svg-url-loader": "^2.3.2",
+    "tapable": "^1.0.0",
     "ts-loader": "^4.0.1",
-    "typescript": "^2.2.1",
+    "typescript": "^2.4.1",
     "url-loader": "^1.0.1",
-    "webpack": "^4.1.0",
+    "webpack": "^4.1.1",
+    "webpack-aspnet-middleware": "^2.2.0",
+    "webpack-bundle-analyzer": "^2.11.1",
     "webpack-cli": "^2.0.12",
-    "webpack-hot-middleware": "^2.12.2",
+    "webpack-hot-middleware": "^2.21.2",
     "webpack-merge": "^4.1.2"
   }
 }
