@@ -16,16 +16,11 @@ namespace flightsearch
     {
         public static void Main(string[] args)
         {
-            //var airports = ImportCsvAsObject("./Model/airports.csv");
-            var host = new WebHostBuilder()
-                .UseKestrel()
-                .UseContentRoot(Directory.GetCurrentDirectory())
-                .UseIISIntegration()
+            BuildWebHost(args).Run();
+        }
+        public static IWebHost BuildWebHost(string[] args) =>
+            WebHost.CreateDefaultBuilder(args)
                 .UseStartup<Startup>()
                 .Build();
-
-            host.Run();
-        }
-
     }
 }
