@@ -1,17 +1,17 @@
 ﻿<!DOCTYPE html>
 <html>
-    <head>
-        <meta charset="utf-8" />
-        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <title>@ViewData["Title"] - Flight Search</title>
+<head>
+	<meta charset="utf-8" />
+	<meta name="viewport" content="width=device-width, initial-scale=1.0" />
+	<title>@ViewData["Title"] - Flight Search</title>
+	<link rel="stylesheet" href="~/dist/bootstrap.css" asp-append-version="true" />
+	<link rel="stylesheet" href="~/dist/vendor.css" asp-append-version="true" />
+	<link rel="stylesheet" href="~/dist/site.css" asp-append-version="true" />
+</head>
+<body>
+	@RenderBody()
 
-        <link rel="stylesheet" href="~/dist/vendor.css" asp-append-version="true" />
-        <link rel="stylesheet" href="~/dist/site.css" asp-append-version="true" />
-    </head>
-    <body>
-        @RenderBody()
-
-        <script src="~/dist/vendor.js" asp-append-version="true"></script>
-        @RenderSection("scripts", required: false)
-    </body>
+	<script src="~/dist/vendor.js" asp-append-version="true"></script>
+	@RenderSection("scripts", required: false)
+</body>
 </html>
